+++ src/Items/Drawer/ItemFormContainer.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { LoginRequired } from "../../../Auth/Login/helpers/LoginRequired";
-import { useItemDrawer } from "../../state/ItemDrawerContext";
+import { LoginRequired } from "../../Auth/Login/helpers/LoginRequired";
+import { useItemDrawer } from "../state/ItemDrawerContext";
 import { ItemForm } from "./ItemForm";
 
 export function ItemFormContainer() {
